+++ src/services/@types/fse.ts
@@ -12,10 +12,8 @@ import { DiffFile } from '../../@types/diffs';
 
 //	Exports ____________________________________________________________________
 
-export type Callback = (error?:null|Error, result?:StatsMap) => void;
-
 export type WalkTreeOptions = {
-	ignore?:string[],
+	excludes?:string[],
 };
 
 export type StatsMap = {